make submissions fetch static
